Rename screens to views. Model View ViewModel not Model Screen ScreenModel.
