+++ lib/main.dart
@@ -1,13 +1,13 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:sp_flutter_app/screens/event_map.dart';
-import 'package:sp_flutter_app/screens/help_screen.dart';
-import 'package:sp_flutter_app/screens/profile_screen.dart';
-import 'package:sp_flutter_app/screens/settings_screen.dart';
-import 'package:sp_flutter_app/screens/wrapper.dart';
-import 'package:sp_flutter_app/shared/constants.dart';
-import 'package:sp_flutter_app/viewmodels/user_viewmodel.dart';
+import 'views/event_map.dart';
+import 'views/help.dart';
+import 'views/profile.dart';
+import 'views/settings.dart';
+import 'views/wrapper.dart';
+import 'shared/constants.dart';
+import 'viewmodels/user_viewmodel.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
