+++ lib/views/views.dart
@@ -1,0 +1,9 @@
+export 'authenticate/authenticate.dart';
+export 'authenticate/sign_in.dart';
+export 'authenticate/register.dart';
+export 'home/home.dart';
+export 'event_map.dart';
+export 'help.dart';
+export 'profile.dart';
+export 'settings.dart';
+export 'wrapper.dart';
